Convert API to using pure functions
Benchmarks showed that pre-computing and using a cache wasn't faster
than just computing on-the-fly. The algorithm came from Python where it
probably made everything a bit faster, so I understand that it was
there.

This also makes it a lot easier to use threads as you don't need to
juggle around reference counting structures between the threads to be
able to re-use the cache.

It will also use a lot less memory since it no longer needs to allocate
memory for all the caches.